feat: exclusionGlobs
Users can provide a comma separated list of glob patterns to have the files matching these patterns be removed / not be brought into Snyk.
For new projects that are not yet in Snyk if they match the glob pattern while cloning the repo, these files will be ignored and not imported.
For files that are already monitored in Snyk, if they match the glob pattern the projects associated with that file (via target file check from the project name) will be deactivated.
This is based on an assumption that anything in exclusions should not be in Snyk full stop