fix: bump node version to latest available 12
Version 12.9.0 used previously had a bug that wouldn't delete folders
recursively. This resulted in cloned repos not being deleted. Upgrading
to the latest available v12 fixes the issue with the exception of
alpine.

Nexe only supports alpine node 12.9.1 which does not have the fix yet.
We will address that in a separate commit, for now unblocking other
builds.

Signed-off-by: Jacek Rzeniewicz <jacek.rzeniewicz@snyk.io>