fix: match project keys exactly in BBS search
The BBS search endpoint matches by prefix, returning repos hosted
in e.g. `bla-1` when searching for just project `bla`.
In some cases this results in two Snyk projects created for the same
manifest in different orgs.

This commit adds a condition to double-check that repos aren't being
added to the wrong org.

This could also be fixed by switching to another API endpoint but prefer
not to make this investment as the project is in maintenance mode now.

Signed-off-by: Jacek Rzeniewicz <jacek.rzeniewicz@snyk.io>