feat: add support for GitHub CloudApp and BitBucket Cloud App for import and sync (#514)
This is a large commit that is primarily focused on adding GitHub CloudApp and BitBucket Cloud App support. A lot of lint, formatting, and dependencies happened with this work that became inseparable. In additionally, the infrastructure for automated testing was no longer available -- so this release was tested manually for each capability.

---------

Signed-off-by: jeff <jeff.brimager@snyk.io>
Co-authored-by: jeff <jeff.brimager@snyk.io>
Co-authored-by: roberthicks <roberthicksiii@gmail.com>
Co-authored-by: Jeff Brimager <211704+sam1el@users.noreply.github.com>